requirement.txt
opencv-python
Pillow
tensorflow
keras
matplotlib